Continue work on UI demo
Clicking on an entity found on the search page now takes you to the
Intake/Questionnaire page. The fields are not populated yet but the
index from the search page is transmitted over so filling those fields
in shouldn't be too bad. Some of the questionnaire is implemented. I got
tired of bootstrap's weirdly spaced out form elements and misbehaving
classes, so I whipped up some simple classes of my own just for element
widths. The form doesn't do anything yet, except for the "Return To
Search" button, which returns you to the search page.