Continue work on UI demo
Remove bootstrap grid - it was only causing problems and this form will
look the same on mobile devices (or will it?).  Add button to add new
client that takes user to questionnaire form only if a search has
returned a negative result. Fix the fake clients' ID numbers so they're
not all the same and do a handful of sex, name, and dob (and age)
changes to make them look more like their pictures (and decrease
unintentional hilarity).