Continue work on UI demo
Rework matching technique to match simultaneously on names and IDs,
which it should have been doing all the time but was buggy until
now. Still a few little bugs with button pisition on the intake form and
clicking "back to results" when adding a new client, but should be
solveable in short order.