Continue work on UI demo
It is now possible to save new clients and modify old ones! (Until you
reload the app of course). Recorded data are very limited. Only the text
input fields are saved. Other values are either hard coded or
ignored. It's just for the demo. Proper fixes will go in later.