Continue work on UI demo
Tried to create a string matching algorithm more flexible than we had
that would allow for greater extensibilty in adding new search terms,
searching for variants on names, and allowing search terms to be entered
in any order. However, there were too many use-cases where correct
behavior was unclear or even in coflict with the proposed
design. Checked in here for posterity. This version should not be
used. The approach to string matching will be resimplified the next time
around.