Continue work on UI demo
In the end, the matching algorithm from last commit wasn't actually that
far from being right. As of this commit it works correctly. I know that
we will not be including "CID" in our matching terms; nor should we
display it on screen. I'm only leaving it here to demonstrate the
capabilitites of the search feature, and demonstrations are more
interesting with more than two matching terms. Given n matching terms
(in this case 3: first name, last name, and CID) an entity will appear
in the results list if the first n-1 whitespace-separated terms entered
by the user are complete matches for any n-1 of the matching terms. The
order in which they are typed doesn't matter. The rightmost term (only)
entered by the user may be a partial match (e.g., if the user enters
"Smith Da" or "David Sm", the entity "David Smith" will appear in the
results, but if the user enters "Sm David" or "Da Smith", it will not).