Continue work on UI demo
Add correct date of birth (DOB) and age functionality. Age in years is
now calculated from the DOB for display in both search results and on
the intake form. DOB is editable on the intake form through the use of a
calendar widget. These date and age functions are enabled through the
use of the Pikaday and Moment libraries. Reduce form fields to those
requested by client. It still remains to make each of these required for
submission. Adjust look of intake form to reflect the reduced number of
fields. Finally, remove "CID" as a searchable datum and remove it from
all displays.