Continue work on UI demo
Allow addtion of new clients even when search returns results, but also
display a warning when adding a new client when results are present. The
warning does not display when no search results are returned. The "add
new client" button continues to be disabled until at least a minimum
number of characters is entered.

Select boxes on the intake form (gender, ethnicity, and race) now
reflect the values of the selected entity, and their values are saved to
that entity if changed by the user.